feat(exports) Create an `ExportKind` int enum in Python.
This patch creates an `ExportKind` enum in Rust. It's layout
representation is `u8`, and the variants are indexed from 1 to 4. This
is important because Python enum's are by default indexed from 1.

An `iter` function returns all the variants of `ExportKind` by index
order.

It is also possible to convert an `ExportKind` variant to a string, to
represent the variant names on the Python side.

Then, to create the enum in Python, we use the [functional
API](https://docs.python.org/3/library/enum.html#functional-api), like
this:

```python
IntEnum("ErrorKind", "FUNCTION MEMORY GLOBAL TABLE")
```

The variant order is defined by `iter`, and the variant name is
defined by the `From` implementation for `&str`.