feat(api) Avoid using a closure to declare host function.
Closures as host functions are going to be disabled for a moment in
Wasmer, see https://github.com/wasmerio/wasmer/pull/1841. In the
`wasmer` Python package, we were passing a closure to
`Function::new_with_env`. In order to get a regular function, this
patch stores the previously captured varibles into the function
environment. The only concerned variable is `result_types`.