feat(api) Avoid cloning `Vec<Type>` for each host function call.
By writing `.map(to_wasm_value)`, we need the types (`Vec<Type>`) to
be owned. The types need to be cloned before the `map` in this
particular context, because they are behind a shared reference.

To avoid cloning the entire `Vec<Type>` everytime a host function is
called (either by the host or by Wasm), we change the `map` to
`.map(|(value, type)| to_wasm_value((value, *ty)))`, which simply
_copy_ the `Type`; thus removing the need to clone `Vec<Type>`, we
keep working on a reference to it.